+++ crates/ruma-client-api/src/appservice/request_ping.rs
@@ -19,8 +19,8 @@ pub mod v1 {
         rate_limited: false,
         authentication: AppserviceToken,
         history: {
-            unstable => "/_matrix/client/unstable/fi.mau.msc2659/appservice/:appservice_id/ping",
-            1.7 => "/_matrix/client/v1/appservice/:appservice_id/ping",
+            unstable("fi.mau.msc2659") => "/_matrix/client/unstable/fi.mau.msc2659/appservice/:appservice_id/ping",
+            1.7 | stable("fi.mau.msc2659.stable") => "/_matrix/client/v1/appservice/:appservice_id/ping",
         }
     };