+++ crates/ruma-client-api/src/authenticated_media/get_media_config.rs
@@ -18,8 +18,8 @@ pub mod v1 {
         rate_limited: true,
         authentication: AccessToken,
         history: {
-            unstable => "/_matrix/client/unstable/org.matrix.msc3916/media/config",
-            1.11 => "/_matrix/client/v1/media/config",
+            unstable("org.matrix.msc3916") => "/_matrix/client/unstable/org.matrix.msc3916/media/config",
+            1.11 | stable("org.matrix.msc3916.stable") => "/_matrix/client/v1/media/config",
         }
     };